+++ src/Ville.java
@@ -5,11 +5,11 @@ public class Ville extends Noeud{
 	 * @param args
 	 */
 	String nom;
-	int x;
-	int y;
+	double x;
+	double y;
 	
 	/**CONSTRUCTEURS*/
-	public Ville(int x, int y, String nom){
+	public Ville(double x, double y, String nom){
 		super();
 		this.x=x;
 		this.y=y;
